requirements.txt
streamlit
pandas
openpyxl
qrcode
reportlab
PyPDF2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib